Some cleanup and comments
